Started to convert register page. Still need to fix form inputs to work with formik.
