+++ meteor-app/imports/startup/client/router.jsx
@@ -23,7 +23,7 @@ import HelpPage from "/imports/ui/pages/help.jsx";
 import LoginPage from "/imports/ui/pages/login";
 import MyAccountPage from "/imports/ui/pages/my-account.jsx";
 import NotFoundPage from "/imports/ui/pages/not-found.jsx";
-import RegisterPage from "/imports/ui/pages/register.jsx";
+import RegisterPage from "/imports/ui/pages/register";
 import UserPage from "/imports/ui/pages/user.jsx";
 
 import CompanyCreateProfileForm from "/imports/ui/pages/create-company-profile";
