+++ meteor-app/imports/ui/pages/register/index.js
@@ -1,0 +1,1 @@
+export default from "./register.jsx";
